+++ emulation/rocjitsu/docs/sphinx/conceptual/architecture.md
@@ -79,7 +79,7 @@ This layer handles loading, decoding, and transforming GPU code objects:
     `code/patch/`).
 -   **Register analysis** --- Backward liveness analysis and def-use
     chains over kernel CFGs provide free-register search used by both
-    DBT and DBI (under `analysis/`).
+    DBT and DBI (under `code/analysis/`).
 
 ### ISA layer
 
@@ -196,7 +196,7 @@ lib/
       code/             Code object loader, basic block analysis
       code/dbt/         Dynamic binary translator
       code/patch/       Code object patcher, spill manager
-      analysis/         Register liveness and def-use analysis
+      code/analysis/    Register liveness and def-use analysis
       config/           JSON/FlatBuffers configuration
   util/                 Shared utilities
   python/amdisa/        ISA codegen pipeline
