+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/code/dbt/semantic/cdna4_to_rdna4.cpp
@@ -4,7 +4,7 @@
 /// @file semantic/cdna4_to_rdna4.cpp
 /// @brief CDNA4-to-RDNA4 handwritten semantic expansion rules.
 
-#include "rocjitsu/analysis/liveness.h"
+#include "rocjitsu/code/analysis/liveness.h"
 #include "rocjitsu/code/builders/instruction_builder.h"
 #include "rocjitsu/code/dbt/hazard_tracker.h"
 #include "rocjitsu/code/dbt/semantic/cdna4_to_rdna_common.h"
