+++ emulation/rocjitsu/tests/patch/spill_manager_test.cpp
@@ -4,8 +4,8 @@
 #include "rocjitsu/code/patch/spill_manager.h"
 
 #include "decode_test_util.h"
-#include "rocjitsu/analysis/exec_state.h"
-#include "rocjitsu/analysis/liveness.h"
+#include "rocjitsu/code/analysis/exec_state.h"
+#include "rocjitsu/code/analysis/liveness.h"
 #include "rocjitsu/code/basic_block.h"
 #include "rocjitsu/code/code_object.h"
 #include "rocjitsu/isa/decoder.h"
