+++ emulation/rocjitsu/tests/smem_sbase_operand_test.cpp
@@ -27,7 +27,7 @@
 ///     not the first source operand.
 
 #include "decode_test_util.h"
-#include "rocjitsu/analysis/def_use_chain.h"
+#include "rocjitsu/code/analysis/def_use_chain.h"
 #include "rocjitsu/code/rj_code.h"
 #include "rocjitsu/isa/decoder.h"
 #include "rocjitsu/isa/instruction.h"
